get_layer_changes/requirements.txt
requests
docker
pyyaml
simplejson